+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/configuration/FeatureConfiguration.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.api;
+package co.elastic.apm.android.sdk.internal.configuration;
 
 public interface FeatureConfiguration {
     boolean isEnabled();
